Config now works at all times.
