+++ src/logs/neoworldbuilder.log
@@ -67,3 +67,5 @@
 {"event": "Configuration loaded successfully", "timestamp": "2025-01-20T21:02:28.041766Z", "level": "info"}
 {"event": "Configuration loaded successfully", "timestamp": "2025-01-20T21:06:47.647632Z", "level": "info"}
 {"event": "Configuration loaded successfully", "timestamp": "2025-01-20T21:10:48.266171Z", "level": "info"}
+{"event": "Configuration loaded successfully", "timestamp": "2025-01-20T21:18:04.848387Z", "level": "info"}
+{"event": "Configuration loaded successfully", "timestamp": "2025-01-20T21:21:06.423015Z", "level": "info"}
